Context has 0 or 1 repository
