+++ src/RepoM.Core.Plugin/Expressions/IRepositoryExpressionEvaluator.cs
@@ -4,9 +4,9 @@ using RepoM.Core.Plugin.Repository;
 
 public interface IRepositoryExpressionEvaluator
 {
-    string EvaluateStringExpression(string value, params IRepository[] repository);
+    string EvaluateStringExpression(string value, IRepository? repository);
 
-    object? EvaluateValueExpression(string value, params IRepository[] repository);
+    object? EvaluateValueExpression(string value, IRepository? repository);
 
     bool EvaluateBooleanExpression(string? value, IRepository? repository);
 }
