+++ tests/RepoM.Plugin.AzureDevOps.Tests/ActionProvider/ActionAzureDevOpsGetPullRequestsV1MapperTests.cs
@@ -42,7 +42,7 @@ public class ActionAzureDevOpsGetPullRequestsV1MapperTests
         _action.ProjectId = "dummy-project-id";
         _action.RepositoryId = null;
         A.CallTo(() => _evaluator.EvaluateBooleanExpression("dummy-Active-property", _repository)).Returns(true);
-        A.CallTo(() => _evaluator.EvaluateStringExpression("dummy-project-id", A<IRepository[]>._)).Returns("real-project-id");
+        A.CallTo(() => _evaluator.EvaluateStringExpression("dummy-project-id", A<IRepository>._)).Returns("real-project-id");
     }
 
     [Fact]
@@ -139,7 +139,7 @@ public class ActionAzureDevOpsGetPullRequestsV1MapperTests
     {
         // arrange
         _action.ProjectId = "dummy-project-id";
-        A.CallTo(() => _evaluator.EvaluateStringExpression("dummy-project-id", A<IRepository[]>._)).Returns(projectId!);
+        A.CallTo(() => _evaluator.EvaluateStringExpression("dummy-project-id", A<IRepository>._)).Returns(projectId!);
 
         // act
         IEnumerable<RepositoryActionBase> result = _sut.Map(_action, _repositories, _composition);
